reworking navigation design and layout
